+++ test/TheWeekend.t.sol
@@ -16,37 +16,22 @@ contract TheWeekendTest is Test, IERC1155Receiver {
         // Test weekend detection for Saturday and Sunday
 
         // Saturday timestamp
-        uint saturdayMorning = 1688803200; // Saturday 8am EST
+        uint256 saturdayMorning = 1688803200; // Saturday 8am EST
         // Sunday timestamp
-        uint sundayEvening = 1688938800; // Sunday 6pm EST
+        uint256 sundayEvening = 1688938800; // Sunday 6pm EST
 
         // Test weekend days
-        assertTrue(
-            theWeekend.isWeekEnd(saturdayMorning),
-            "Saturday should be weekend"
-        );
-        assertTrue(
-            theWeekend.isWeekEnd(sundayEvening),
-            "Sunday should be weekend"
-        );
+        assertTrue(theWeekend.isWeekEnd(saturdayMorning), "Saturday should be weekend");
+        assertTrue(theWeekend.isWeekEnd(sundayEvening), "Sunday should be weekend");
 
         // Test non-weekend days
-        uint mondayMorning = 1689001200; // Monday 8am EST
-        uint thursdayNoon = 1688652000; // Thursday 12pm EST
-        uint fridayNoon = 1688738400; // Friday 12pm EST
+        uint256 mondayMorning = 1689001200; // Monday 8am EST
+        uint256 thursdayNoon = 1688652000; // Thursday 12pm EST
+        uint256 fridayNoon = 1688738400; // Friday 12pm EST
 
-        assertFalse(
-            theWeekend.isWeekEnd(mondayMorning),
-            "Monday should not be weekend"
-        );
-        assertFalse(
-            theWeekend.isWeekEnd(thursdayNoon),
-            "Thursday should not be weekend"
-        );
-        assertFalse(
-            theWeekend.isWeekEnd(fridayNoon),
-            "Friday should not be weekend"
-        );
+        assertFalse(theWeekend.isWeekEnd(mondayMorning), "Monday should not be weekend");
+        assertFalse(theWeekend.isWeekEnd(thursdayNoon), "Thursday should not be weekend");
+        assertFalse(theWeekend.isWeekEnd(fridayNoon), "Friday should not be weekend");
     }
 
     function testFuzz_DayOfWeek(uint256 timestamp) public view {
@@ -74,29 +59,25 @@ contract TheWeekendTest is Test, IERC1155Receiver {
     }
 
     // IERC1155Receiver implementation
-    function onERC1155Received(
-        address,
-        address,
-        uint256,
-        uint256,
-        bytes calldata
-    ) external pure override returns (bytes4) {
+    function onERC1155Received(address, address, uint256, uint256, bytes calldata)
+        external
+        pure
+        override
+        returns (bytes4)
+    {
         return this.onERC1155Received.selector;
     }
 
-    function onERC1155BatchReceived(
-        address,
-        address,
-        uint256[] calldata,
-        uint256[] calldata,
-        bytes calldata
-    ) external pure override returns (bytes4) {
+    function onERC1155BatchReceived(address, address, uint256[] calldata, uint256[] calldata, bytes calldata)
+        external
+        pure
+        override
+        returns (bytes4)
+    {
         return this.onERC1155BatchReceived.selector;
     }
 
-    function supportsInterface(
-        bytes4 interfaceId
-    ) external pure override returns (bool) {
+    function supportsInterface(bytes4 interfaceId) external pure override returns (bool) {
         return interfaceId == type(IERC1155Receiver).interfaceId;
     }
 }
